change Sickle main entry point to ts.SourceFile
ts.Program (the previous parameter) is a fully compiled program of
potentially multiple files, while ts.SourceFile is a single parsed file.
It appears from how we're using sickle that we want the latter, not the
former.

(It may be the case that we need a fully compiled program to get all the
references right, but for now this appears to work.)

Closes #13.