handle some more cases of subclassing needing super() calls
When we generate our own constructor() for a class (which we do
when we need to declare any properties on it), we must also put
in a call to super().  But this is easy to get wrong so do our
best to avoid calling super() when possible.

Helps with issue #38 but doesn't close it.