fix enum export by upgrading closure compiler
If you wrote "export enum ...", the generated ES6 JavaScript would
fail to compile in the Closure compiler.  I tracked this down to
a bug that was fixed in the upstream Closure compiler, so to fix
this I had to make us use a newer version of the upstream compiler.

To use a newer version of the upstream compiler, I removed one of
our indirection layers: we previously used the 'closure-compiler'
npm library which pulled in 'google-closure-compiler', but now the
latter module exposes a simple API so it's easy enough to just use
it directly.

Fixes issue #50.