skip comments when exporting enums
When we export an enum we try to prepend an @typedef on it.
If the enum already had a comment on it, the comment interferes
with the type annotation.  Fix this by stripping comments when
emitting enums.  See the diff of es6/enum.js to see where this
is necessary.

This comment stripping is pretty hacky and will likely fail in
complex scenarios.  Right now it's limited to enums so maybe it
will work out?

See issue #52 for an example of why we can't do something obvious
to skip the comment -- the parse tree is pretty complex.