use a helper instead of a synthetic constructor
We need Closure to believe that all our member variables exist.
To do so, we need to tag them with an /** @type */ bit.  But it
turns out you don't need to do this in the constructor, it can
be done anywhere!

Doing this annotation in a unused extra method means that we
don't need to create constructors, which means we don't need
to worry about calling super() correctly, completely avoiding
issue #38.

Because the extra method is unused, Closure ought to eliminate
it when you compile, too, so it costs nothing.