make the type annotations helper static
For fields on a class, we emit a helper function that mentions
all the fields for Closure's sake.  But the helper function cannot
be a method, as that affects the type of the generated class (it
has the additional method).

Fix this by making the helper static.
See the new "structural.untyped" test for a verification of this fix.