parse jsdoc to handle @param/@return properly
When emitting a function, we want to preserve any JSDoc describing
the parameters but supply our own values for the types.

TypeScript has a JSDoc parser, but it's not exposed in a place
where we can use it.  So for now do a hacky regexp-based parse
of JSDoc.  The interfaces used here are pretty similar to the
TypeScript ones so if the API does expose them, it shouldn't be
too hard to adjust.

Fixes issue #67.