convert types to closure types
Summary:
In the normal traversal of the program's AST, when we encounter
a TypeScript type (e.g. "any") we want to emit "any" because we're
still emitting TypeScript.

But in the cases where we want to generate a Closure type (e.g.
for a docstring) we want to emit "?" for "any".  So in those cases
use a separate recursive traversal.

Reviewers: rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D20