split Annotator into two classes
Summary:
The superclass "Rewriter" manages the AST traversal and node emitting,
while Annotator then implements a node-visit function and handles all
the same logic it did before.

This split is so we can use the Rewriter logic a second time, in the
module rewriter (the piece that processes JS->JS to inject goog.module
etc.).

More work to prepare for #111.

Reviewers: rkirov, mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D41