handle "export *" in module rewriter
Summary:
Sickle already rewrites "export *" for Closure's benefit
before it ever gets to the TypeScript transpile. But in the
devserver case we don't use Sickle on the inputs, so we still
need to handle the TS output for "export *", which looks like:
  __export(require(...))

Because this only happens on the non-Closure codepath, we can
translate it to just
  __export(goog.require(...))
(with the appropriate new module name).

Reviewers: alexeagle

Reviewed By: alexeagle

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D47