always emit goog.module, even in empty-looking files
Summary:
The bug was that we'd look for the first statement before producing
a goog.module, so if the file had no statements we wouldn't produce one.

The code looked like it was trying to be clever and emit the goog.module
after an initial file comment, but it didn't actually work.  Do the simpler
equivalent thing for now.

More work on #111.

Reviewers: alexeagle

Reviewed By: alexeagle

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D58