don't crash on generic arrow functions in typed mode
Summary:
When attempting to generate the type annotations for code like:
  function generic<T>(callback: (T) => void);
  generic(foo => {});
at the point where we generate the "foo => {}" bit, when producing
a type for the arrow function, the type passed in by TypeScript is
funky in a way I can't figure out how to generate.

It turns out we don't generate types on arrow functions anyway in
untyped mode, so just broaden that to also happen in typed mode.

Reviewers: mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D75