run goog.module processor on test suite, fix enums
Summary:
Running the test suite without the goog.module postprocessor meant that
we were generating code not acccepted by the Closure compiler in some
cases.  Fix this by running convertCommonJsToGoogModule in the test suite.

This revealed that the way enums were implemented didn't work when the
enums were exported, due to weirdness in Closure for aliasing typedefs.
For now just emit them as plain numbers; again, we've already type checked
so it's ok to use a more relaxed type.

Reviewers: mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D82