feat: include the file path as module.id in the generated source.
Summary:
The module.id is set to the path of the source file being passed in, which
should be the project relative path of the source as seen by the TypeScript
compiler. Users should take care not to pass in absolute paths.

This allows TypeScript files to e.g. refer to templates relative to their own
source location.

Reviewers: evmar

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D133