don't treat interfaces as values when emitting decorators
Summary:
In the decorator-munger, if a ctor has a param
  foo: SomeInterface
we cannot use SomeInterface as a value in the metadata.

To determine whether SomeInterface is a value or a type, we need
to type-check the program.  This means changing the API to require
a ts.Program.

To make the tests work against this new API, all tests needed to
be changed to be type-checkable.  But refactorings elsewhere in
tsickle meant that we can now reuse the program creation logic from
test_support, which deletes a bunch of code.

That test support code is also more careful about caching lib.d.ts
so we can also now always type-check the result of the decorator
emit in the tests, so I also deleted the conditional compilation
check.

Fixes #148.

Reviewers: mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D139