Correctly compare paths to files in the current directory.
If the tsickle main function is run against files in the same directory,
then closure type comments are not added to the outputted JavaScript. For
example, if you run tsickle against './hello.ts' then the outputted
hello.js will not have type comments.

This occurs because the sourceReplacingCompilerHost compares fileNames to
determine whether we ran tsickle on the source file. 'hello.ts' is compared
to './hello.ts' and we incorrectly skip the file.

This is fixed by comparing the path to the files using the TypeScript
sys utilities.