handle default imports in the import rewriter
Summary:
The input code
  import Foo from 'bar';
  console.log(Foo);
becomes
  var mod = require('bar');
  console.log(mod.default);
from TypeScript compilation, so if Foo is a type we need to alias it
in case it appears in any JSDoc.

To do this, I factored the default import branch to share code with
the named import (e.g. "import {a as b}") branch.  They're handled in
a pretty similar manner.

Reviewers: rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D143