assert that inputs have been typechecked
Summary:
Throw an exception if you pass in a file that isn't from a ts.Program.

It turns out that if if you accidentally pass in a plain ts.SourceFile,
many TS APIs will still work but then others will fail with a very
mysterious exception that wastes days of our time.  :P

Reviewers: alexeagle

Reviewed By: alexeagle

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D149