provide a type for forward declared variable
Summary:
Recall that if another module exports a type (e.g. an interface),
tsickle will create a Closure @record variable from it in the other
module.  When this module imports that type, for Closure purposes
we need to declare a reference to the *value* of that record variable.

But at TS compilation time that @record doesn't yet exist, so we must
forward declare it using "declare".  This change gives that forward
declaration a type.  By the time Closure gets to compilations using
this variable, this declaration has entirely disappeared; it's just
to make the TS compiler happy in the interim.

More work on #112.

Reviewers: rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D153