make decorator-annotator not eat jsx + extend test suite
Summary:
Extend the golden test suite to also run decorator annotation.
If the decorator-annotator changes a program, we save the changed
program with a golden filename like "foo.decorated.ts".

Extend a golden test to show that the above works.

Fix JSX mangling by moving the JSX bug workaround to a place that
is shared by both the main tsickle processing and the decorator
annotation.

Fixes #167.

Reviewers: alexeagle, rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D170