decorator-annotator: preserve whitespace
Summary:
When code relies on ASI, then preserving the whitespace
attributed to a decorator node is necessary to preserve the
functioning of the program.

Fixes #169.

(I think I should add some API that just preserves newlines, not spaces,
but I will do that in a followup because this current pattern is all
over tsickle.)

Reviewers: alexeagle, rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D172