don't overide typescript source argument, mkdirp before write (#171)
Currently Tsickle will always transpile the whole program every time it
is run. This pull request adds support for transpiling only the files
supplied to ts via the command line. For example tsickle -- foobar.ts
would now only transpile foobar.ts instead of the whole program.

This also calls mkdirp on the directory path of a file before it is
written. Previously, if an output directory was given to the ts
transpiler via the command line (i.e. tsickle -- --outDir build/)
tsickle would be unsuccessful in writing the file if the entire
directory tree had not been created at the specified output path.