treat jsdoc problems as warnings, not errors
Summary:
If there's a problem in a file's jsdoc, return it to the caller as
a warning, not as an error.  This will allow the downstream
consumer of tsickle to decide whether to be strict (reject any
code with warnings) or lax (allow).

In particular, this is necessary to compile third-party code where
the user cannot fix the code to avoid warnings.

Part of issue #177.

Reviewers: mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D195