use ts.forEachChild rather than node.getChildren()
Summary:
It turns out that node.getChildren() actually re-parses the AST
of the Node given to it and it produces sometimes-wrong parses.

ts.forEachChild is a higher-level function that iterates through
the semantically-meaningful children of a Node, based on its
member variables.  So when you iterate a "class" for example you
no longer iterate through the curly braces as separate children,
but rather only the higher-level children like properties.

By using this function we work around the existing JSX emit bug,
and it works around a TS 2.0 problem (issue #183).  It's also
simpler and likely faster than the old code because it doesn't
reparse and it visits fewer nodes.

Reviewers: mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D198