treat multiple ctors in d.ts as just a debug warning
Summary:
If a d.ts specifies multiple ctors for the same
class, we currently only produce the first and ignore the
rest.  Rather than aborting compilation, instead treat it
as a debug warning.

(This will allow us to treat errors as errors in third-party
code; currently we ignore both errors and warnings. I am
trying to make us obey errors.)

Reviewers: rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D204