stop typechecking tsickle output
Summary:
We need to be able to generate invalid TypeScript.

Drop typechecking of tsickle output from the test suite.
Note that we still Closure-compile the tsickle output so the
output is still checked; it's just that sometimes we're gonna
want to do things that TypeScript disallows, like reach past
a "private" modifier, to make Closure work.

See issue #189.

Reviewers: rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D214