properly export interface functions
Summary:
A TypeScript interface becomes a Closure function.
The previous code attempted to export that function when the
interface was exported, but due to
  https://github.com/Microsoft/TypeScript/issues/10122
it actually wasn't exporting.

You can see this in the new test goldens, where export.js now
includes "Bar" as an export.  (Bar is a TypeScript interface but
at the Closure level it is a value and must be in the module
exports.)

This also lets us remove a crazy forward declare hack that actually
wasn't necessary -- it was just working around that the export
wasn't working.

This matches the exporting behavior in d090fe41885ddc994e6774c688cf90682740008c.

Reviewers: rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D213