drop the tsickle_Foo renaming hack
Summary:
Now that we no longer typecheck, we can use a much simpler
pattern for local import renaming.

Recall that the problem is that a user can write:
  import Foo from 'bar';
  /** @type {Foo} */
  var x = new Foo();
which after TS compilation becomes:
  var bar_1 = require('bar');
  /** @type {Foo} */
  var x = new bar_1.Foo();

This is problematic because the JSDoc no longer points at an imported
name.

The new fix in this change is to add, in TS, a line like:
  const Foo = Foo;
after the import.  This looks crazy but it actually compiles into
  const Foo = bar_1.Foo;
which makes a type alias on the Closure side.

It's much simpler than the old hacks where we rename a bunch of stuff!

Reviewers: rkirov

Reviewed By: rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D217