generate module.id in ES6 mode that can be optimized out by closure. (#192)
The current module.id generation is careful to preserve the global module if it's already present. However in that form closure cannot optomize it out even if module.id is not referenced in the file - Leading to the final bundle bloat.

The fix is to preserve the existing module only in ES5 mode and just output module.id in ES6 mode(where we don't expect module to already exist).