Special-case construct signatures.
Indicating the Closure that this is a constructor is probably more important
than specifying all constructor fields. As a side effect, this also translates
types specified in the `var`/`interface` style into the type expected by
Closure:

    var Document = {
      new(): Document;
      prototype: Document;
    };
    interface Document { ... }
    let x = Document;

Now translates to just:

    let /** @type {function(new: Document): ?} */ x = Document;

Which is accepted by newer Closure versions and unblocks migration.