Move tsickle type annotations outside of the class definition.
TypeScript will rename references to a class within its body if it is
decorated, so that the code inside refers to the decorated version of the
class:

    @Decorator
    class X { foo() { X.prototype.whatever; } }

    var X_1 = class X { foo() { X_1.prototype.whatever; } }

For closure compiler, X_1.prototype.whatever does not define a field on the
class X (it doesn't understand the alias), and so type checking fails later on.

This change moves the field declarations outside of the class definition to
avoid the issue:

    @Decorator
    class X { ... }
    /** @type {...} */ X.prototype.whatever;

Which then gets converted to:

    let X = class X_1 { ... }
    /** @type {...} */ X.prototype.whatever;

Which makes Closure happy.