Hide property declaring code in a function.
It turns out that a statement like "Foo.prototype.bar" might actually execute code if "bar" on Foo.prototype is a getter. That code then fails because its this pointer is not set in the invocation, breaking load of tsickle coverted code. Some users have an odd build setup where they execute tsickle-converted code in the browser, so we need to hide these statements in a function to avoid breaking code on load.