Assign a default export in goog.module()s
Closure Compiler works on improving strictness for goog.module(), in particular enforcing that named exports are immutable. The exports = {} statement indicates to Closure Compiler that there is a default export, so Closure Compiler does not error on mutating/accessing named exports on it.

This is more of a workaround than a fix, but TypeScript users are guaranteed to only ever access named exports through indirection off of the module export object (that is, TypeScript compiler makes sure to implement mutable exports on the usage site), so for TS users the error that Closure Compiler tries to prevent here does not actually occur.