handle "this" parameters
When a function has a parameter with the special name "this", that
is specifying the type of the "this" object in the body of the function
and shouldn't show up in the parameter list.

It turns out that the issue in #280 was actually caused by us getting
confused about the function's parameter list.  Properly handling the
'this' argument lets us recognize that the parameter is actually a
...rest argument so fixing the 'this' issue fixes that as well.

Fixes #274 and #280!