initial ts2.1 support
TypeScript 2.1 changed some APIs (which was easy enough to handle)
but it appears the representation of types is pretty different, so
I'm leaving that for a later change.

This change is enough to get tsickle to *compile* with TypeScript 2.1.
A bunch of type-related code is commented out (marked with "TODO(ts2.1)")
and many of the tests fail.

Initial work on #295.