ts2.1: rearrange type translation, use new ObjectFlags
Make the main "translate" function match the enum in the
TypeScript API.  Split out the handling of more complex
Object types into a separate function, and adjust the code
there to use the new objectFlags field.

This actually makes most tests pass(!) but there are still
a few more to track down, and there are still a bunch of
TODOs.

More work on #295.