be careful around type aliases in @implements (#336)
It's legal to @implements a type alias, but we need to be careful
to not @implements a type alias that maps back to a symbol that
we didn't emit a type for due to the symbol being both a type
and a value.  Yikes!

To ensure I got this right, I added both a test for:
- implementing an ordinary interface via an alias
- a more complex reproduction of the zone issue

Fixes #333 harder.