special-case optional any properties
When a property is optional, e.g.
  foo?: string;
Then the type of the property is string|undefined and the above
translation handles it correctly.
But in the special case of any, which includes undefined:
  foo?: any;
The type is just "any" (because that covers undefined as well) and
our translation of the type is just "?".  But Closure wants it to
instead be ?|undefined.

Fixes #287.