Rewrite tsickle's aliasing to use goog.forwardDeclare.
TypeScript emits imports with unpredictable module prefixes. Because
tsickle needs to refer to them in comments, it used to create local
aliases:

    var Foo = Foo; // -> compiles to: var Foo = some_module_prefix_1.Foo

This causes problems:
* the aliases confuse Closure Compiler into thinking the module code is
  actually used in a value position, which causes warnings for cyclical
  imports.
* this forces module load statements (i.e. goog.require) to be emitted
  for modules that would otherwise not be loaded by TS, as they're only
  used in type positions. This causes load order and cyclical import
  issues.
* they don't work for non-value types, in particular type aliases.
  Interfaces work mostly by accident, because tsickle emits function
  values for them.

To fix, change our code to introduce explicit local aliases for symbols
by emitting a synthetic `goog.forwardDeclare`. `goog.forwardDeclare` has
the semantics of telling Closure Compiler a type exists, without forcing
a runtime load of the corresponding code, which matches the semantics of
a type-only import in TypeScript (i.e. no runtime load).

This requires introducing the aliases on import, and then using a lookup
table for symbols when emitting types.

Because several tools in the Closure toolchain purge files that are
never required (e.g. `--closure_entry_point`), only using
`forwardDeclare` can lead to a failure in type checking, as the file
declaring types might have been removed.

To avoid that issue at least in the most common case, this code
explicitly emits a `goog.require` for modules that only declare types,
which then causes them to be loaded again.

Overall, this should emit code for type references that is more
Closure-idiomatic.

Additionally, this change emits aliases on the export object for type
aliases. Type aliases are different from all other exports because they
do not introduce a value in Closure, so they need to be emitted as a
pure property access with an appropriate aliasing comment.

TESTED=unittests, and compiling an Angular 2 example with type checks