Fix error when handling input with some files with preexisting source maps, and some without. (#392)
When composing source maps, we assumed that if any of the source files had a source map for a given transform, they all had them.  While this is correct for decorator downleveling and closurization, we don't have any guarantees for inline source maps in the source files, thus we need to guard against the source map not existing.