Retain @fileoverview comments preceding non-value imports.
TypeScript drops imports from the output if none of the imported symbols
are ever used as values (as opposed to types). In Closure, those
comments can have semantic meaning attached to them. Make sure the
comment gets retained by splitting it off the import, through a single
line break between the comment and the import.