always emit interfaces, even in untyped mode
This changes inches us closer to fully typed mode, by making us
always emit a type definition even in untyped mode.

We accomplish this by moving around a few if statements.  The
resulting interfaces don't have heritance clauses (that broke
in google3) and the fields don't have types, but it means that
at least a type that is declared in TypeScript is visible (albeit
slightly mangled) to Closure code.