Emit abstract methods as pure declarations in a helper function.
This should help Uglify and similar simple optimizers to remove the declarations entirely.

I verified that the emit is actually checked by Closure - it does understand abstract function declarations in this form.

This also fixes some lint hinting to use "const" instead of "let", and the more modern cast form.