Support namespaces (aka modules) in ambient exports.
This allows users to write code like:

    export declare namespace myNamespace { ... }

This retains the existing behaviour that tsickle moves ambient
declarations into the global namespace, i.e. for the above code a global
Closure namespace "myNamespace" will be created. This is required at
least for namespaces that contain values - otherwise Closure would
reject code accessing the values. We might need to reconsider and have
to mangle the namespace for pure value definitions if we find this
causes collisions. That means "declare global ..." and "declare
namespace" have the same effect in tsickle.

Special care must be taken when types in ambients refer to non-ambient
types. Closure does not have a syntax for externs types to refer to
types in modules, so we emit these as the unknown type.

This change also avoids some problems with "namespace" usage. While
namespaces generally emit code that Closure might not understand,
tsickle will now at least avoid emitting types referring to namespaces.