avoid redeclaring augmented modules
Assume that the first person who declares a given namespace
will emit the externs for it, so we shouldn't emit them again.

In the output externs, you can see that the "local" namespace
is still emitted, but we avoid re-emitting this "angular"
namespace because the module we're augmenting already emitted
it.