Always emit enums as proper Closure types.
This was previously avoided to keep the output valid TypeScript wise,
but that is not necessary, as we do not type check the intermediate
output.

This also changes the TypeScript->JS emit to include Closure types for
const enums that disappear in runtime emit. This slightly deviates from
TS semantics, as const enums now create runtime values. The values
should be inlined by Closure Compiler though, and we specifically do not
emit the name -> value mapping.

While testing, this unconvered an unrelated issue in our emit: as we
emit TypeScript enum types as `type Foo = number;`, TypeScript will
complain if it is instructed to generate a .d.ts and the code contains
exported literal types including the enum value (`export const x:
Foo.BAR;`). This was previously the case, with this CL just adding a
test for it.

This is in turn symptom of a larger problem: we cannot reliably emit
.d.ts from tsickle-processed source code, as all kinds of things might
go wrong. The fix is, for the time being, to not use tsickle to emit
.d.ts, and in the long term move to transformer mode (which is a good
idea anyway).

Fixes #535.