Ensure TypeScript can emit a .d.ts for the given code.
If the effective type shape of a program refers to private names,
TypeScript cannot generate a .d.ts for it, and fails during emit.
Previously tsickle did not test for this.

This requires updating fields.ts. Because it is not a module, it
effectively exports every global variable it defines, which means its
private class type is affected by this change. Simply changing it to a
module resolves the problem.