Emit just `?` for templatized unhandled types.
tsickle needs to emit `?` for types in several situations. If such a
type is templatized, e.g. `T<U>`, tsickle would previously translate `T`
to `?`, and then emit `?<U>`, which is illegal syntax in Closure
Compiler.

This change avoids the problem by just returning `?` directly in such
situations.

This also adds a test for blacklisted parameterized types which I
originally suspected to be the trigger. It turns out the problem was
caused by type/value conflicts, but the test coverage still seems
useful, so I'm leaving these in.