Only skip emitting text for comments.
The goal of the `skipUpToOffset` mechanism is to avoid emitting comments
that have already been processed. Commit b298d950 accidentally changed
this to skip emitting any text up to `skipUpToOffset`, by nesting the
skipping operation.

This meant that emitting nodes that preceded the
start offset of the current node in the source file would silently not
emit any text, which caused regression #547.

This commit changes the behaviour to only apply skipping for emit starts
that are within the trivia range of the given node, i.e. between
`getFullStart` and `getStart`. That achieves the goal of skipping
already emitted comments without accidentally ever skipping "real"
emitted text.