Refactor tsickle golden tests.
Instead of expressing the different test results as patches inline,
rather store completely separate files for the variants (regular,
transformer, and onepass). This makes it easier to review the variants
in isolation.

This change also moves diagnostics for transformer mode into a separate
file. As the full transformation runs in one go, it is difficult to
associate diagnostics with specific files, and there are only .js files
around, so we cannot use a custom format to store diagnostics inline
without potentially disturbing Closure Compiler.

Additionally, this change seeks to reduce the duplication between
transformer and regular golden tests a bit by using the same golden
comparisons, and also updates .vscode/launch.json to match the new
locations.