Provide a source mapping for every emit.
When tsickle rewrites a source file, it produces text 2 ways.  First it copies text from an AST node to the output.  Second, it produces its own text and outputs it.  Previously, we only provided source mappings for the first case, since we had the node to map back to.  This PR changes rewriter.ts to keep track of the current node on a stack as it traverses the AST and uses that node to provide source mappings for the second kind of emit.  In certain cases the classes that extend rewriter.ts don't use its node traversal mechanisms to visit the children of certain node types it transforms - in those cases the node stack is manually updated.

This PR also adds a bunch of source map tests, roughly one for each kind of AST node that we transform.  In writing those tests 2 issues arrose:

1.  Typescript doesn't produce good source maps for multiline/jsdoc comments, so we can't currently provide source maps for function signature annotations.
2.  The transformer system doesn't like having custom emits mapped back to original nodes - I got a bunch of errors initially and I was forced to preserve the existing source mapping behavior for the transformer system.  I'll have to work with Tobias on this.