Don’t get confused with comments on import statements that are elided:
Due to a bug in TypeScript, we have to store comments for import/export
statements in a separate data structure
and reapply them after the TypeScript transformer is done.

The old algorithm matched the import/export statements by index
and therefore got confused when an import/export statement was elided.

The new algorithm matches by location in the original source file.