adjust enum test to be closer to reality
With a single-element enum, in code like:
  let x: MyEnum.VALUE = MyEnum.VALUE;
the type of x is just 'MyEnum'.  This might be a bug in TS
but it's certainly the case that this test isn't trying to
exercise that corner of the language.  Avoid it.

This is more preliminary work for TypeScript 2.4.
(Same as 170bb99, but in another file.)