Handle @fileoverview in all initial comments.
The fileoverview comment transformer handled @fileoverview comments that weren't attached to any
node, but @fileoverview comments just have to be before any js, so they can be in leading comments on the first statement in the file.  Thus, we need to look for @fileoverview in a leading NotEmittedStatement's comments as well as in the leading comments on the first emitted statement.  To avoid handling comments that were on a node that isn't emitted (ie an interface), move the fileoverview transformer to run after tsc runs its compile.