Enable writing new goldens. (#625)
Previously, updating goldens worked because the runfiles directory contains symlinks back to the original file.  However, if we needed to create a new golden file, there wasn't a symlink, so we'd create the new golden in the bazel runfiles directory, not in the test_files directory.  By parsing the symlink to the source ts files for the test, we can find the absolute path to the tsickle workspace root, which allows us to construct the path to the actual file we want to update.