Upgrade tsickle to TS2.5.3 (#627)
The major change is that NodeArray is now readonly, so we copy it when
we need to modify.

Also, getEmitFlags is gone from the public TS api, but still present at
runtime. Waiting to hear from TS team about better fix for that.

Some todos left for better cleanups - tracked by #634.