Fix closure_externs.js.
Due to e2e_test not correctly executing, a number of errors have crept
in. The recurring pattern is that Closure does not actually allow
`@typedef`s of generic types, and thus silently ignored the template
arguments passed. This change rephrases the given types as `@record`
(where possible) or `@constructor` (where necessary) of the aliased
types, which allows `@template` parameters to be passed.