Run Closure using NTI and OTI.
It turns out there are some patterns that break in OTI (like using
generic types on local variables) and some that break in NTI (like using
generic arguments on typedefs). So we'll need to run both, and hope
there's a subset between those two type checkers that'll work for us.