Only resolve aliases when emitting typedefs.
Previously, tsickle would not emit type alias, but rather always their
resolved underyling type. E.g. for a `type X = A|B;`, tsickle would emit
`A|B` in all places where `X` was referenced as a type. This largely
works, except if the constituent types of the type alias are actually
not in scope at the emit location, e.g. because they are not imported.

This change introduces two emit modes for aliases: the default mode,
where aliases should be emitted as their symbol directly, via
`type.aliasedSymbol`. This is used everywhere, including in locations
where code introduces a new `@typedef` aliasing an existing type.

The second mode resolves the alias, i.e. it emits the underlying type
instead of the alias' symbol. This is used in the two locations that
actually define a new `typedef` - the writer for regular type aliases,
and the externs writer.

This also required slightly tweaking the code that prevents recursive
emits of types to check its re-entrance condition in the right location.