Fix golden_tsickle_test to check all diagnostics.
Since revision 71391b6b36 (by yours truly), our test runner would only
check for warnings, but not for error diagnostics anymore.

We did not notice the change in the massive update to the various test
files included, and apparently tslint also does not warn for unused
variables in destructuring assignments.