Forward declare `import * as x from ...` imports.
tsickle previously forward-declared symbols from named and default
imports (`import {x} from ...`, `import X from ...`), but did not
forward declare symbols from prefix imports (`import * as x ...`). This
causes Closure errors if the prefix import is elided by TypeScript emit
for imports that are only used in type positions.

Previously, the main motivation of the forward declares was to avoid
issues with the module prefixing being unpredictable, which is not a
problem with prefix imports. However a secondary benefit of forward
declarations is that even if TypeScript elides the import for pure type
imports, the forward declare is still present for Closure. This change
harmonizes the handling of imports in this regard.

As a side note, it seems quite hard to detect if an import will be
elided by tsc or not. The relevant APIs are internal, and even hackily
accessing them fails due to the type checker being in different modes
at different stages. In either case, a forward declare is the correct
translation in Closure Compiler terms.