forward declare symbols not in the local scope.
TypeScript's type checker expands aliases in union types as they are
constructed. This is apparently a design limitation and hard to change
for reasons including performance.

In tsickle, the expanded symbols might however not be available at the
use site, e.g. if a union type is declared in `fileA.ts`, contains
symbols available locally, but then imported in `fileB.ts` where its
constituent types are not available:

    // fileA
    export class LocalType {}
    export type TypeDef = LocalType|number;

    // fileB
    import {TypeDef} from './fileA';
    let useTypeDef: TypeDef|string;

Tsickle would emit `LocalType|number|string` as the type for
`useTypeDef`. This change fixes the issue by ensuring all symbols that
are used have an appropriate forward declare. This turns out to be
easier than mapping back into the original type or trying to emit the
syntactical type.