Allow offsetting the transformer source map.
The tranformer source map is used to map newly created nodes back to
pre-transformation original nodes, and required for a correct TypeScript
emit (e.g. mapping symbols to their imported identifiers).

Adding support for an offset allows code to prepend additional text to
the generated output *after* emit has happened (and thus after mappings
have been established). This is useful to insert additional required
imports that are only discovered while processing the file.

This is a bit of a hack, but relatively localized, and should go away
if/when we move to a proper TS transformer.

This commit also makes a number of cleanups and clarifications in the
source map handling code:
- remove optionality markers in locations where a value is always passed
- move DefaultSourceMapper to the one place that uses it and rename it
  It is never used outside of testing.
- Add some documentation on what's going on in transformer_sourcemap.