Separate forward declares with a line break.
Previously, if a list of imports would end in a trailing force-load of a
type only module, the line comment would effectively strip the next line
of code. This is very rare (people separate their imports and code), but
gets triggered when forward declaring a symbol from a type only module
which gets prepended before all other imports.