Treat const enums like types for imports.
TypeScript inlines all values of a `const enum` during compilation (if
`preserveConstEnums` is `false`), so that there will never be runtime
imports of the symbol. To avoid the same problems as with modules
defining only types, `const enum`s must therefore be treated like types.

This only affects the import situation, other judgements based on
whether a symbol is a value (such as generating an export) still hold.